refactor: fix name
